changes to dicom reader to take into account different slopes and patients with multiple scan trials

+++ application/stage1.py
@@ -132,7 +132,7 @@ class Stage1DataLoader(StandardDataLoader):
                 continue
 
             if "target" in tags:
-                sample[OUTPUT][tag] = np.int64(self.labels[set][sample_index])
+                sample[OUTPUT][tag] = np.int32(self.labels[set][sample_index])
 
             if "sample_id" in tags:
                 sample[OUTPUT][tag] = sample_id
